Anim: when evaluating animation layer, add special case for strip overlap
Add a special case to the animation layer evaluation, for when the
evaluation time is exactly on the last frame of an animation strip. In
this case that strip should be used, unless it is also overlapping with
another strip, in which case that strip is used.

Example: there are two strips `A=[t1, t2]` and `B=[t3, t4]`, and they
are evaluated at time `t2`.

- When `t2 < t3`, only strip `A` should be evaluated.
- When `t2 >= t3`, only strip `B` should be evaluated.

The current code doesn't mix between strips yet, so for now this is
sufficient.