Anim: hook up the Animation data-block to property state colors in the UI
Expand `BKE_animadata_fcurve_find_by_rna_path()` so that it can also find
animation data stored in an `Animation` data-block.

If there is a Keyframe Strip overlapping the given frame time, and it
has a matching FCurve, that will be returned. Otherwise it will fall back
to searching Keyframe Strips that do not overlap. This at least ensures
a consistent "is this property animated?" indicator, even when it may not
be animated at this particular point in time.

Note that this might be a bit dangerous, as I didn't fully investigate all
(indirect) uses of the FCurve returned by this function.