Anim: make `Output::name` a proper name
`Output::name` is now unique within the `Animation` data-block. This also
changes slightly what happens when assigning an animated ID to another
Output, or when un-assigning an ID from the Animation.

Assigning an ID to an Output will rename the Output to that ID, but ONLY
if the Output has no name yet. Otherwise the name will just be kept as-is.

Unassigning an Animation from an ID will write back the last used Output
name to the ID's AnimData. This is basically a workaround for the fact that
this doesn't happen on Output rename yet. That is marked as TODO in the
code for future implementation.